pyproject.toml
[project]
name = "invoice-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.41.0",
    "ipykernel>=6.30.1",
    "pdfplumber>=0.11.7",
]
